+++ nion/ui/PyQtProxy.py
@@ -402,6 +402,7 @@ class PyRadioButton(QtWidgets.QRadioButton):
     def __init__(self):
         super().__init__()
         self.object = None
+        self.setAutoExclusive(False)
         self.clicked.connect(self.__clicked)
 
     def __clicked(self):